+++ README.md
@@ -1,4 +1,4 @@
-# BEM Helper
+# reBEM Helper
 
 [![CircleCI](https://circleci.com/gh/remithomas/bem-helper.svg?style=svg)](https://circleci.com/gh/remithomas/bem-helper)
 
@@ -7,6 +7,12 @@ Work In Progress but still A simple BEM helper
 ## How to install
 
 `...`
+
+## Development
+
+```bash
+yarn start
+```
 
 ## How to use
