+++ __tests__/BemHelper_test.js
@@ -3,23 +3,23 @@
 
 var Jest = require("@glennsl/bs-jest/src/jest.js");
 var Block = require("bs-platform/lib/js/block.js");
-var BemHelper = require("../src/BemHelper.js");
+var ReBemHelper = require("../src/ReBemHelper.js");
 
 describe("BEM helper", (function () {
         Jest.test("it should create a className for block", (function () {
-                var myBemClassNames = BemHelper.bem(/* Some */["my-block"], /* None */0, /* None */0, /* () */0);
+                var myBemClassNames = ReBemHelper.bem(/* Some */["my-block"], /* None */0, /* None */0, /* () */0);
                 return Jest.Expect[/* toBe */2]("my-block", Jest.Expect[/* expect */0](myBemClassNames));
               }));
         Jest.test("it should create a className for element", (function () {
-                var myBemClassNames = BemHelper.bem(/* None */0, /* Some */["my-element"], /* None */0, /* () */0);
+                var myBemClassNames = ReBemHelper.bem(/* None */0, /* Some */["my-element"], /* None */0, /* () */0);
                 return Jest.Expect[/* toBe */2]("my-element", Jest.Expect[/* expect */0](myBemClassNames));
               }));
         Jest.test("it should create classNames for block and element", (function () {
-                var myBemClassNames = BemHelper.bem(/* Some */["my-block"], /* Some */["element"], /* None */0, /* () */0);
+                var myBemClassNames = ReBemHelper.bem(/* Some */["my-block"], /* Some */["element"], /* None */0, /* () */0);
                 return Jest.Expect[/* toBe */2]("my-block__element", Jest.Expect[/* expect */0](myBemClassNames));
               }));
         Jest.test("it should create classNames for block and modifier with truthly boolean modifier", (function () {
-                var myBemClassNames = BemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
+                var myBemClassNames = ReBemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
                         /* Boolean */Block.__(1, [
                             "active",
                             /* true */1
@@ -29,7 +29,7 @@ describe("BEM helper", (function () {
                 return Jest.Expect[/* toBe */2]("my-block my-block--active", Jest.Expect[/* expect */0](myBemClassNames));
               }));
         Jest.test("it should create classNames for block and modifier with falsy boolean modifier", (function () {
-                var myBemClassNames = BemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
+                var myBemClassNames = ReBemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
                         /* Boolean */Block.__(1, [
                             "active",
                             /* false */0
@@ -39,14 +39,14 @@ describe("BEM helper", (function () {
                 return Jest.Expect[/* toBe */2]("my-block", Jest.Expect[/* expect */0](myBemClassNames));
               }));
         Jest.test("it should create classNames for block and modifier with string modifier", (function () {
-                var myBemClassNames = BemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
+                var myBemClassNames = ReBemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
                         /* String */Block.__(0, ["big"]),
                         /* [] */0
                       ]], /* () */0);
                 return Jest.Expect[/* toBe */2]("my-block my-block--big", Jest.Expect[/* expect */0](myBemClassNames));
               }));
         Jest.test("it should create classNames for block and modifier with string modifier and boolean modifier", (function () {
-                var myBemClassNames = BemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
+                var myBemClassNames = ReBemHelper.bem(/* Some */["my-block"], /* None */0, /* Some */[/* :: */[
                         /* String */Block.__(0, ["big"]),
                         /* :: */[
                           /* Boolean */Block.__(1, [
@@ -59,7 +59,7 @@ describe("BEM helper", (function () {
                 return Jest.Expect[/* toBe */2]("my-block my-block--big my-block--active", Jest.Expect[/* expect */0](myBemClassNames));
               }));
         return Jest.test("it should create classNames for block element and modifier with string modifier and boolean modifier", (function () {
-                      var myBemClassNames = BemHelper.bem(/* Some */["my-block"], /* Some */["element"], /* Some */[/* :: */[
+                      var myBemClassNames = ReBemHelper.bem(/* Some */["my-block"], /* Some */["element"], /* Some */[/* :: */[
                               /* String */Block.__(0, ["big"]),
                               /* :: */[
                                 /* Boolean */Block.__(1, [
